requirements.txt
streamlit==1.36.0
   PyMuPDF==1.24.1
   python-docx==1.1.2
   spacy==3.7.2
   sentence-transformers==2.7.0
   google-generativeai==0.8.3
   pandas==2.2.2
   https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl